+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 *[CalVer, YY.month.patch](https://calver.org/)*
 
 ## Future
+- TRIO107 now completely ignores any function whose body consists solely of ellipsis, pass, or string constants.
 - TRIO103, 107 and 108 now inspects `while` conditions and `for` iterables to avoid false alarms on a couple cases where the loop body is guaranteed to run at least once.
 
 ## 22.8.6